Refactor: variables 위치 수정
