+++ src/api/log/log.service.ts
@@ -1,72 +1,37 @@
 import { Injectable } from '@nestjs/common';
-import { InjectRepository } from '@nestjs/typeorm';
-import { Connection, Repository } from 'typeorm';
-import { PAGE_SIZE, Sort, Status } from '../../common/variables.util';
+import { DataSource, Repository } from 'typeorm';
+import { PAGE_SIZE, Sort, Status } from './entity/variables.util';
 import { Log } from './entity/log.entity';
 
 @Injectable()
 export class LogService {
-  constructor(
-    @InjectRepository(Log)
-    private readonly projectRepository: Repository<Log>,
-    private connection: Connection,
-  ) {}
+  constructor(private dataSource: DataSource) {}
   async retrieveUsers(request: any) {
-    const queryRunner = this.connection.createQueryRunner();
-    await queryRunner.connect();
-    await queryRunner.startTransaction();
     try {
-      let users = [];
-      const requesterCount = await this.projectRepository.createQueryBuilder(
-        Log,
-        'log',
-      );
+      let logs = [];
 
       // query 생성
-      const queryResult = await getManager().createQueryBuilder(
-        GridgeRequester,
-        'requester',
-      );
+      const queryResult = this.dataSource.createQueryBuilder(Log, 'log');
 
       // 상태 필터
       if (request.query.status == Status.ACTIVE) {
-        await queryResult.andWhere('requester.status IN (:status)', {
-          status: Status.ACTIVE,
-        });
-
-        await requesterCount.andWhere('requester.status IN (:status)', {
+        await queryResult.andWhere('log.status IN (:status)', {
           status: Status.ACTIVE,
         });
       } else if (request.query.status == Status.INACTIVE) {
-        await queryResult.andWhere('requester.status IN (:status)', {
-          status: Status.INACTIVE,
-        });
-
-        await requesterCount.andWhere('requester.status IN (:status)', {
+        await queryResult.andWhere('log.status IN (:status)', {
           status: Status.INACTIVE,
         });
       }
 
-      // 날짜순 내림차순, 오름차순 정렬, 닉네임 가나다 순 정렬
+      // 날짜순 내림차순, 오름차순 정렬
       if (request.query.sort == Sort.DESC) {
-        await queryResult.orderBy('requester.createdAt', 'DESC');
-        await requesterCount.orderBy('requester.createdAt', 'DESC');
+        await queryResult.orderBy('log.createdAt', 'DESC');
       } else if (request.query.sort == Sort.ASC) {
-        await queryResult.orderBy('requester.createdAt', 'ASC');
-        await requesterCount.orderBy('requester.createdAt', 'ASC');
-      } else if (request.query.sort == 'realName') {
-        await queryResult.orderBy('requester.realName', 'ASC');
-        await requesterCount.orderBy('requester.realName', 'ASC');
-      } // 현재 GridgeRequester DB에는 nickname 칼럼이 없다.
+        await queryResult.orderBy('log.createdAt', 'ASC');
+      }
 
-      await queryResult.select([
-        'requester.id',
-        'requester.realName',
-        'requester.email',
-        'requester.phoneNumber',
-        'requester.createdAt',
-        'requester.status',
-      ]);
+      await queryResult.select(['log.id', 'log.createdAt', 'log.status']);
 
       let page = 0;
       // 페이징;
@@ -77,26 +42,20 @@ export class LogService {
         } else {
           page = (request.query.page - 1) * PAGE_SIZE;
         }
-        // await queryResult.offset(start).limit(pageSize);
         await queryResult.take(PAGE_SIZE).skip(page);
       }
 
       // 조회
-      users = await queryResult.getMany();
-
-      const counting = await requesterCount.select('requester.id').getMany();
+      logs = await queryResult.getMany();
 
       const data = {
-        userCount: counting.length,
-        user: users,
+        log: logs,
       };
 
-      // const result = makeResponse(response.SUCCESS, data);
-
-      // return result;
+      return data;
     } catch (error) {
       console.log(error);
-      // return response.ERROR;
+      return error;
     }
   }
 }
