+++ src/app.module.ts
@@ -4,9 +4,14 @@ import { AppService } from './app.service';
 import { UserModule } from './api/user/user.module';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { typeOrmAsyncModuleOptions } from './config/typeorm.config';
+import { LogModule } from './api/log/log.module';
 
 @Module({
-  imports: [UserModule, TypeOrmModule.forRootAsync(typeOrmAsyncModuleOptions)],
+  imports: [
+    UserModule,
+    TypeOrmModule.forRootAsync(typeOrmAsyncModuleOptions),
+    LogModule,
+  ],
   controllers: [AppController],
   providers: [AppService],
 })
